// src/lib/common/types.ts
export type AnyType = any;
/**
 * Allowed data type for JTDSchema
 */
export type AllowedType =
  | 'string'
  | 'boolean'
  | 'timestamp'
  | 'int8'
  | 'uint8'
  | 'int16'
  | 'uint16'
  | 'int32'
  | 'uint32'
  | 'float32'
  | 'float64';

export type ParsedMeta<T> = O.Record<string, T>;

/**
 * Type representation for unfinished JTD Schema built from decorators
 *
 * When `definitions` contains no more class, a schema will be treated as a normal JTDSchema and will be used to generate a `serializer` / `parser`
 */
export type ParsedJTDSchema = {
  properties: Record<string, ParsedTypeDefinition>;
  definitions: Record<string, C.Class | O.Object>;
};

export type ParsedTypeDefinition = {
  ref?: string;
  type?: string;
  metadata?: { transform?: Transform };
  elements?: {
    type?: string;
    ref?: string;
  };
  nullable?: boolean;
  validateNested?: boolean;
};

export type CompiledTransform = (o: O.Object) => string;

export type CompileOption = {
  strict?: boolean;
};

export type ErrorMessageTemplate = {
  toString(property: string, value: unknown): string;
};

export type Constraint = {
  errorMessage?: O.Record<string, string | ErrorMessageTemplate['toString']>;
} & O.Object;

export type BaseTypeDef = {
  nullable?: boolean;
  isArray?: boolean;
};

export type PrimitiveType = {
  type: AllowedType;
} & BaseTypeDef;

export type ClassType = {
  class: C.Class;
  validateNested?: boolean;
} & BaseTypeDef;

export type TypeDef = PrimitiveType | ClassType;

type Transform = <V, O, C>(value: unknown, object: O, context?: C) => V;

export type Format = {
  name: string;
  type: string | number;
  validate(value: unknown): boolean;
  /**
   * Automatically rename if the given format name is already used
   * @default false
   */
  autoRename?: boolean;

  readonly newName?: string;
};

export type Meta = {
  constraint?: Constraint;
  typeDef?: TypeDef;
  transform?: Transform;
  allowEmpty?: boolean;
  format?: Format;
};

export type InsertMeta<TMeta extends Meta> = O.AtLeast<O.Required<TMeta>>;

export type MergedMeta<TMeta extends Meta> = Omit<TMeta, 'format'> & {
  formats: Format[];
};

export type ParsedResult = {
  constraints: ParsedMeta<Constraint>;
  types: ParsedJTDSchema;
  allowNullables: [string[], string[]];
  formats: Format[];
};

export type CompiledResult = {
  serialize: (o: O.Object) => string;
  parse: JTDParser;
  validate: CtxValidatorFunction;
};

export interface IYieldResultCallback {
  (error: Error): void;
  (error: undefined, data: CompiledResult): void;
}
export type ParseMetaOptions = {
  /**
   * Allow overriding message for a single validation or a group of validations on the property
   *
   * If `true`, register `ajv-error` package to the ajv instance
   *
   * @see https://github.com/ajv-validator/ajv-errors
   *
   * @default false
   */
  useCustomMessage?: boolean;

  /**
   * Validate nested object or an array of objects
   *
   * @default false
   */
  validateNested?: boolean;

  registeredFormats?: string[];
};

export type TypedAjvStorageOption<TParsedResult extends ParsedResult> =
  ParseMetaOptions & {
    /**
     * Compile each schema in a single tick by `setImmediate` api
     *
     * @default false
     */
    compileAsync?: boolean;

    allErrors?: boolean;

    constructJSONSchemaFns?: ConstructJSONSchema<TParsedResult>[];

    constructJTDSchemaFns?: ConstructJTDSchema<TParsedResult>[];

    /**
     * Extends Ajv instance
     */
    ajv?: (ajv: Ajv) => void;
  };
// ...
export type ConstructJSONSchema<TParsedResult extends ParsedResult> = (
  schema: JSONSchemaType<AnyType>,
  parsedMeta: TParsedResult,
  source: O.Object
) => JSONSchemaType<AnyType>;

export type ConstructJTDSchema<TParsedResult extends ParsedResult> = (
  schema: ParsedJTDSchema,
  parsedMeta: TParsedResult,
  source: O.Object
) => ParsedJTDSchema;

// export type ConstructJTDType<TP>

export interface CtxValidatorFunction<T = unknown>
  extends Pick<
    ValidateFunction,
    'errors' | 'evaluated' | 'schema' | 'schemaEnv' | 'source'
  > {
  (data: unknown, context?: unknown): data is T;
}